+++ packages/vue/tests/Link.spec.ts
@@ -101,7 +101,7 @@ describe("<curi-link>", () => {
       expect(a.textContent).toBe("Kokomo");
     });
 
-    it("spreads forward object onto anchor", () => {
+    it("spreads additonal props onto anchor", () => {
       wrapper = new Vue({
         el: node,
         template: `
@@ -109,7 +109,7 @@ describe("<curi-link>", () => {
             <curi-link
               name="Place"
               :params="{ name: 'Aruba' }"
-              :forward="{ class: 'hooray' }"
+              class="hooray"
             >
               Aruba
             </curi-link>