+++ packages/vue/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Remove `forward` prop from link components.
+
 ## 2.0.0-alpha.4
 
 * Use `router.url`.